requirements.txt
annotated-types==0.7.0
anyio==4.9.0
certifi==2025.6.15
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
cryptography==45.0.5
fastapi==0.115.14
greenlet==3.2.3
h11==0.16.0
idna==3.10
python-multipart==0.0.20
pyyaml==6.0.2
pycparser==2.22
pydantic==2.11.7
pydantic-core==2.33.2
pymysql==1.1.1
requests==2.32.4
sniffio==1.3.1
sqlalchemy==2.0.41
starlette==0.46.2
typing-extensions==4.14.1
typing-inspection==0.4.1
urllib3==2.5.0
uvicorn==0.35.0
Pillow==11.1.0
ultralytics==8.3.100
prometheus-fastapi-instrumentator